+++ main.tf
@@ -39,9 +39,24 @@ locals {
 
   # If provided, set Dataflow worker to new user-managed service account;
   # otherwise, use Compute Engine default service account
-  dataflow_worker_service_account = ((var.dataflow_worker_service_account != "")
-    ? google_service_account.dataflow_worker_service_account[0].email
-  : "${data.google_project.project.number}-compute@developer.gserviceaccount.com")
+  dataflow_worker_service_account = ((var.dataflow_worker_service_account != "") ?
+    ((var.use_externally_managed_dataflow_sa == false) ?
+      google_service_account.dataflow_worker_service_account[0].email :
+    var.dataflow_worker_service_account) :
+  "${data.google_project.project.number}-compute@developer.gserviceaccount.com")
+
+  # Cross validate Dataflow service account variables
+  # TODO: Remove once https://github.com/hashicorp/terraform/issues/25609 is resolved or use pre-conditions if version > 1.2.0
+  # tflint-ignore: terraform_unused_declarations
+  validate_sa_settings = (var.use_externally_managed_dataflow_sa && var.dataflow_worker_service_account == "") ? tobool("Please pass dataflow_worker_service_account if you want to manage it externally.") : true
+  # Grant Dataflow worker service account required IAM roles over external resources
+  # **unless** using an externally managed worker service account (Option 1 and 2):
+  grant_service_account_roles = var.use_externally_managed_dataflow_sa == false
+  # Grant caller (you!) permissions to impersonate service account
+  # **only** when creating a new worker service account (Option 2)
+  grant_service_account_impersonation = (
+    var.dataflow_worker_service_account != "" && var.use_externally_managed_dataflow_sa == false
+  )
 
   subnet_name           = coalesce(var.subnet, "${var.network}-${var.region}")
   project_log_sink_name = "${var.dataflow_job_name}-project-log-sink"